[fix]setResult 공통 부분 추상메서드로 추출
