+++ src/main/java/com/youtil/Common/Handler/AbstractAiRequestHandler.java
@@ -1,8 +1,10 @@
 package com.youtil.Common.Handler;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.youtil.Common.Constants.AiServiceConstants;
 import com.youtil.Concurrency.RedisSemaphoreManager;
+import io.jsonwebtoken.io.SerializationException;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.PriorityBlockingQueue;
@@ -12,6 +14,7 @@ import java.util.concurrent.TimeUnit;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.data.redis.RedisConnectionFailureException;
 import org.springframework.data.redis.connection.stream.MapRecord;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.web.reactive.function.client.WebClientRequestException;
@@ -146,13 +149,34 @@ public abstract class AbstractAiRequestHandler<T, Q> {
         acknowledgeAndDelete(record);
     }
 
+    protected void setErrorResult(String requestId) {
+
+        try {
+            redisTemplate.opsForValue().set(
+                    constants.getResultKey() + requestId,
+                    objectMapper.writeValueAsString(getEmptyErrorResponse()),
+                    constants.getResultTtl());
+        } catch (JsonProcessingException e) {
+            log.error("에러 응답 저장 실패", e);
+        } catch (RedisConnectionFailureException e) {
+            log.error("레디스 접속 에러", e);
+        } catch (SerializationException e) {
+            log.error("직력화 실패", e);
+
+        } catch (IllegalArgumentException e) {
+            log.error("부적절한 값 포함", e);
+        } catch (Exception e) {
+            log.error("알수없는 예외가 발생했습니다.", e);
+        }
+    }
+
     protected abstract String getAiType();
 
     protected abstract T handleRequest(String requestJson, long userId) throws Exception;
 
     protected abstract void logSuccess(String requestId);
 
-    protected abstract void setErrorResult(String requestId);
+    protected abstract Object getEmptyErrorResponse();
 
     protected abstract Q wrap(MapRecord<String, Object, Object> record);
 }
